move parsers into dir
