+++ tdrs-backend/tdpservice/parsers/factory.py
@@ -1,7 +1,8 @@
 """Factory class for all parser classes."""
 
-from tdpservice.parsers.fra_parser import FRAParser
-from tdpservice.parsers.tdr_parser import TanfDataReportParser
+from tdpservice.parsers.parser_classes.fra_parser import FRAParser
+from tdpservice.parsers.parser_classes.program_audit_parser import ProgramAuditParser
+from tdpservice.parsers.parser_classes.tdr_parser import TanfDataReportParser
 
 
 class ParserFactory:
@@ -15,6 +16,8 @@ class ParserFactory:
                 return TanfDataReportParser
             case "FRA":
                 return FRAParser
+            case "Program Audit":
+                return ProgramAuditParser
             case _:
                 raise ValueError(
                     f"No parser available for program type: {program_type}."
@@ -24,12 +27,15 @@ class ParserFactory:
     def get_instance(cls, **kwargs):
         """Construct parser instance with the given kwargs."""
         program_type = kwargs.pop("program_type", None)
-        match program_type:
-            case "TAN" | "SSP":
-                return TanfDataReportParser(**kwargs)
-            case "FRA":
-                return FRAParser(**kwargs)
-            case _:
-                raise ValueError(
-                    f"No parser available for program type: {program_type}."
-                )
+        return cls.get_class(program_type)(**kwargs)
+        # match program_type:
+        #     case "TAN" | "SSP":
+        #         return TanfDataReportParser(**kwargs)
+        #     case "FRA":
+        #         return FRAParser(**kwargs)
+        #     case "Program Audit":
+        #         return ProgramAuditParser(**kwargs)
+        #     case _:
+        #         raise ValueError(
+        #             f"No parser available for program type: {program_type}."
+        #         )
